Replaced longs.
git-svn-id: svn://svn.code.sf.net/p/opensludge/code@111 27a91dd4-ec73-4ff0-838b-64dfa05f697e

+++ source/Engine/savedata.cpp
@@ -1,3 +1,5 @@
+#include <stdint.h>
+
 #include <string.h>
 
 #include "ALLFILES.H"
@@ -63,7 +65,7 @@ char * readStringEncoded (FILE * fp) {
 }
 
 char * readTextPlain (FILE * fp) {
-	long startPos;
+	int32_t startPos;
 
 	int stringSize = 0;
 	bool keepGoing = true;
